+++ spec/ruby_rabbitmq_janus/process/concurrencies/event_admin_spec.rb
@@ -3,8 +3,7 @@
 require 'spec_helper'
 
 describe RubyRabbitmqJanus::Process::Concurrencies::EventAdmin, type: :thread,
-                                                                name: :event_admin,
-                                                                retry: 3 do
+                                                                name: :event_admin do
   let(:publish_name) { :pub_admin }
   let(:listener) { RubyRabbitmqJanus::Rabbit::Listener::FromAdmin }
   let(:exception_runner) { RubyRabbitmqJanus::Errors::Process::EventAdmin::Run }