+++ spec/ruby_rabbitmq_janus/process/concurrencies/event_spec.rb
@@ -3,8 +3,7 @@
 require 'spec_helper'
 
 describe RubyRabbitmqJanus::Process::Concurrencies::Event, type: :thread,
-                                                           name: :event,
-                                                           retry: 3 do
+                                                           name: :event do
   let(:publish_name) { :pub_classic }
   let(:listener) { RubyRabbitmqJanus::Rabbit::Listener::From }
   let(:exception_runner) { RubyRabbitmqJanus::Errors::Process::Event::Run }